fixed no authentication on delete franchise endpoint

+++ src/routes/franchiseRouter.js
@@ -11,7 +11,17 @@ franchiseRouter.docs = [
     path: '/api/franchise?page=0&limit=10&name=*',
     description: 'List all the franchises',
     example: `curl localhost:3000/api/franchise&page=0&limit=10&name=pizzaPocket`,
-    response: { franchises: [{ id: 1, name: 'pizzaPocket', admins: [{ id: 4, name: 'pizza franchisee', email: 'f@jwt.com' }], stores: [{ id: 1, name: 'SLC', totalRevenue: 0 }] }], more: true },
+    response: {
+      franchises: [
+        {
+          id: 1,
+          name: 'pizzaPocket',
+          admins: [{ id: 4, name: 'pizza franchisee', email: 'f@jwt.com' }],
+          stores: [{ id: 1, name: 'SLC', totalRevenue: 0 }],
+        },
+      ],
+      more: true,
+    },
   },
   {
     method: 'GET',
@@ -19,7 +29,14 @@ franchiseRouter.docs = [
     requiresAuth: true,
     description: `List a user's franchises`,
     example: `curl localhost:3000/api/franchise/4  -H 'Authorization: Bearer tttttt'`,
-    response: [{ id: 2, name: 'pizzaPocket', admins: [{ id: 4, name: 'pizza franchisee', email: 'f@jwt.com' }], stores: [{ id: 4, name: 'SLC', totalRevenue: 0 }] }],
+    response: [
+      {
+        id: 2,
+        name: 'pizzaPocket',
+        admins: [{ id: 4, name: 'pizza franchisee', email: 'f@jwt.com' }],
+        stores: [{ id: 4, name: 'SLC', totalRevenue: 0 }],
+      },
+    ],
   },
   {
     method: 'POST',
@@ -27,7 +44,11 @@ franchiseRouter.docs = [
     requiresAuth: true,
     description: 'Create a new franchise',
     example: `curl -X POST localhost:3000/api/franchise -H 'Content-Type: application/json' -H 'Authorization: Bearer tttttt' -d '{"name": "pizzaPocket", "admins": [{"email": "f@jwt.com"}]}'`,
-    response: { name: 'pizzaPocket', admins: [{ email: 'f@jwt.com', id: 4, name: 'pizza franchisee' }], id: 1 },
+    response: {
+      name: 'pizzaPocket',
+      admins: [{ email: 'f@jwt.com', id: 4, name: 'pizza franchisee' }],
+      id: 1,
+    },
   },
   {
     method: 'DELETE',
@@ -59,9 +80,14 @@ franchiseRouter.docs = [
 franchiseRouter.get(
   '/',
   asyncHandler(async (req, res) => {
-    const [franchises, more] = await DB.getFranchises(req.user, req.query.page, req.query.limit, req.query.name);
+    const [franchises, more] = await DB.getFranchises(
+      req.user,
+      req.query.page,
+      req.query.limit,
+      req.query.name,
+    );
     res.json({ franchises, more });
-  })
+  }),
 );
 
 // getUserFranchises
@@ -76,7 +102,7 @@ franchiseRouter.get(
     }
 
     res.json(result);
-  })
+  }),
 );
 
 // createFranchise
@@ -90,17 +116,22 @@ franchiseRouter.post(
 
     const franchise = req.body;
     res.send(await DB.createFranchise(franchise));
-  })
+  }),
 );
 
 // deleteFranchise
 franchiseRouter.delete(
   '/:franchiseId',
+  authRouter.authenticateToken,
   asyncHandler(async (req, res) => {
+    if (!req.user.isRole(Role.Admin) && !req.user.isRole(Role.Franchisee)) {
+      throw new StatusCodeError('unable to delete a franchise', 403);
+    }
+
     const franchiseId = Number(req.params.franchiseId);
     await DB.deleteFranchise(franchiseId);
     res.json({ message: 'franchise deleted' });
-  })
+  }),
 );
 
 // createStore
@@ -110,12 +141,16 @@ franchiseRouter.post(
   asyncHandler(async (req, res) => {
     const franchiseId = Number(req.params.franchiseId);
     const franchise = await DB.getFranchise({ id: franchiseId });
-    if (!franchise || (!req.user.isRole(Role.Admin) && !franchise.admins.some((admin) => admin.id === req.user.id))) {
+    if (
+      !franchise ||
+      (!req.user.isRole(Role.Admin) &&
+        !franchise.admins.some((admin) => admin.id === req.user.id))
+    ) {
       throw new StatusCodeError('unable to create a store', 403);
     }
 
     res.send(await DB.createStore(franchise.id, req.body));
-  })
+  }),
 );
 
 // deleteStore
@@ -125,14 +160,18 @@ franchiseRouter.delete(
   asyncHandler(async (req, res) => {
     const franchiseId = Number(req.params.franchiseId);
     const franchise = await DB.getFranchise({ id: franchiseId });
-    if (!franchise || (!req.user.isRole(Role.Admin) && !franchise.admins.some((admin) => admin.id === req.user.id))) {
+    if (
+      !franchise ||
+      (!req.user.isRole(Role.Admin) &&
+        !franchise.admins.some((admin) => admin.id === req.user.id))
+    ) {
       throw new StatusCodeError('unable to delete a store', 403);
     }
 
     const storeId = Number(req.params.storeId);
     await DB.deleteStore(franchiseId, storeId);
     res.json({ message: 'store deleted' });
-  })
+  }),
 );
 
 module.exports = franchiseRouter;
